Fix a number of bugs in STM32-FSMC-ICE40-SDRAM project:
 - Fix silly verilog mistakes causing bad register sizes and wrong signals.
 - Fix reversed polarity on sdram_rwn signal.
 - Fix hang caused by sdram-controller missing sdram_adv signal.
 - Work-around hang caused by lost sdram_write_done signal.
 - Fix sdram_gpio pin mappings.

Some things are starting to work now. But there are still many
glitches or bit errors in the SDRAM output. The suspected cause is
LEDs on address pins A[10] and A[3:0] (effectively 1k ohm pulldowns)
and too-high operating speed at 108 MHz.

Also, the lost sdram_write_done signal, which causes a hang without
the work-around, would be nice to investigate and understand.